requirements.txt
beautifulsoup4==4.11.1
htmldate==1.3.2
pandas==1.5.1
python_dateutil==2.8.2
requests==2.28.1
streamlit==1.15.2
openpyxl==3.0.10
xlsxwriter